Improvements for #5 and some logging

+++ src/resxProvider.ts
@@ -2,18 +2,21 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import resx from 'resx';
 import { getNonce } from './utilities/getNonce';
+import { printChannelOutput } from './extension';
 
 export class ResxProvider implements vscode.CustomTextEditorProvider {
 
   public static register(context: vscode.ExtensionContext): vscode.Disposable {
     const provider = new ResxProvider(context);
     const providerRegistration = vscode.window.registerCustomEditorProvider(ResxProvider.viewType, provider);
+    printChannelOutput("ResX Editor custom editor provider registered.", true);    
     return providerRegistration;
   }
 
   private static readonly viewType = 'resx-editor.editor';
   private registered = false;
-
+  private currentPanel: vscode.WebviewPanel | undefined = undefined;
+  
   constructor(
     private readonly context: vscode.ExtensionContext
   ) { }
@@ -23,20 +26,32 @@ export class ResxProvider implements vscode.CustomTextEditorProvider {
     webviewPanel: vscode.WebviewPanel,
     _token: vscode.CancellationToken
   ): Promise<void> {
+    this.currentPanel = webviewPanel;
     webviewPanel.webview.options = {
       enableScripts: true,
       localResourceRoots: [vscode.Uri.joinPath(this.context.extensionUri, 'out')]
     };
     webviewPanel.webview.html = this._getWebviewContent(webviewPanel.webview);
 
-    let disposable = vscode.commands.registerCommand('resx-editor.deleteResource', () => {
+    try
+    {
+    if (!this.registered) {
+      printChannelOutput("deleteResource command registered", true);
+      this.registered = true;
+      let disposable = vscode.commands.registerCommand('resx-editor.deleteResource', () => {
 
-      webviewPanel.webview.postMessage({
-        type: 'delete'
+        this.currentPanel?.webview.postMessage({
+          type: 'delete'
+        });
       });
-    });
 
-    this.context.subscriptions.push(disposable);
+      this.context.subscriptions.push(disposable);
+    }
+  }
+  catch (e)
+  {
+    console.log(e);
+  }
 
     async function updateWebview() {
       webviewPanel.webview.postMessage({
@@ -59,6 +74,9 @@ export class ResxProvider implements vscode.CustomTextEditorProvider {
       switch (e.type) {
         case 'update':
           this.updateTextDocument(document, e.json);
+          return;
+        case 'log':
+          printChannelOutput(e.message, true);
           return;
       }
     });
@@ -87,6 +105,10 @@ export class ResxProvider implements vscode.CustomTextEditorProvider {
                 <head>
                   <meta charset="UTF-8">
                   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+                  <meta
+                    http-equiv="Content-Security-Policy"
+                    content="default-src 'none'; img-src ${webview.cspSource} https:; script-src ${webview.cspSource}; style-src ${webview.cspSource};script-src 'nonce-${nonce}';"
+                  />
                 </head>
                 <body>
                   <vscode-data-grid id="resource-table" aria-label="Basic" generate-header="sticky" aria-label="Sticky Header"></vscode-data-grid>
